KC-1034 Cannot copy PD document - receive incident report

+++ coeus-impl/src/main/java/org/kuali/coeus/propdev/impl/attachment/Narrative.java
@@ -666,7 +666,8 @@ public class Narrative extends KcPersistableBusinessObjectBase implements Hierar
 
     @PostRemove
     public void removeData() {
-        if (getNarrativeAttachment().getFileDataId() != null) {
+        // KC-1034 Cannot copy PD document - receive incident report
+        if (getNarrativeAttachment() != null && getNarrativeAttachment().getFileDataId() != null) {
             getKcAttachmentDao().removeData(getNarrativeAttachment().getFileDataId());
         }
     }
